% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_CNV_call}
\alias{example_CNV_call}
\title{roxygen documentation here}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 21 rows and 4 columns.
}
\usage{
example_CNV_call
}
\description{
some stuff
}
\keyword{datasets}
