% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function1.R
\name{function1}
\alias{function1}
\title{Calculates Information Criteria}
\usage{
function1(loglikelihood, nClusters, dimensionality, observations, probability)
}
\arguments{
\item{loglikelihood}{explanation}

\item{nClusters}{A positive integer indicating the number of clusters}

\item{dimensionality}{A positive integer indicating the dimensionality
of dataset}

\item{observations}{A positive integer indicating the number of observations}

\item{probability}{A vector indicating the probability of each cluster. The
vector should sum to 1.

@return Returns an S3 object of class InfCriteria with results.
\itemize{
\item BICresults - A value of class "numeric" indicating BIC value
\item AICresults - A value of class "numeric" indicating AIC value
\item ICLresults - A value of class "numeric" indicating ICL value
}}
}
\description{
A function that calculates information criteria given log-likelihood,
number of clusters, dimension of dataset, number of observations,
and the probability
}
