% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCNVgenes.R
\name{getCNVgenes}
\alias{getCNVgenes}
\title{Get genes contained within CNVs}
\usage{
getCNVgenes(
  CNV_call,
  chromosome_number = NULL,
  reference_genome = "GRCh38",
  show_piechart = FALSE
)
}
\arguments{
\item{CNV_call}{A dataframe of CNVs containing 4 columns: the chromosome number
on which it is found, the start position, the end position, and
the type (either DUP or DEL).}

\item{chromosome_number}{A string value indicating the chromosome name
on which to identify genes contained within the CNVs.
Can be 1-22, 'X', or 'Y'.}

\item{reference_genome}{A string value, either 'GRCh37' or 'GRCh38', indicating
the genome to compare. Default is 'GRCh38'.}

\item{show_piechart}{A boolean value indicating whether to show the distribution
of genic vs. non-genic CNVs.}
}
\value{
Returns a character vector of genes that are contained within
the input list of CNVs.
}
\description{
A function that produces a list of genes that are contained within Copy Number
Variants(CNVs) from the provided table of CNVs containing the chromosome number,
start position, end position, and type (duplication or deletion). User can also
specify the chromosome number for which they want results, and the reference
genome they want to use to find genes within the CNVs. Optionally, the function
outputs a graph showing the distribution of genic vs. non-genic CNVs.
}
\examples{
# Example 1
# Using sample_CNV_call dataset available with package
# Get list of genes
gene_list <- getCNVgenes(CNV_call = sample_CNV_call)


# Example 2
# Produces piechart plot
gene_list2 <- getCNVgenes(CNV_call = sample_CNV_call, show_piechart = TRUE)

# Example 3
# Specifying chromosome number and reference genome
# Get list of genes for chromosome 22 using the GRCh37 reference genome
gene_list3 <- getCNVgenes(CNV_call = sample_CNV_call,
                          chromosome_number = "22",
                          reference_genome = "GRCh37")


\dontrun{
# Example 4
# Larger dataset, runs slower
large_gene_list <- getCNVgenes(CNV_call = large_CNV_call)
}

}
\references{
Wickham H, François R, Henry L, Müller K, Vaughan D (2023). \emph{dplyr: A Grammar of Data Manipulation}. R package version 1.1.3, \href{https://CRAN.R-project.org/package=dplyr}{Link}.

Durinck S, Spellman PT, Birney E, Huber W (2009). \emph{Mapping identifiers for the integration of genomic datasets with the R/Bioconductor package biomaRt}. Nature Protocols 4, 1184-1191 .

BiomaRt Bioconductor R package documentation. (2023). Retrieved November 13, 2023, from \href{http://useast.ensembl.org/info/data/biomart/biomart_r_package.html}{Link}

Gurbich T, Ilinsky V (2020). \emph{ClassifyCNV: A tool for clinical annotation of copy-number variants}. Scientific Reports, 10(1), Article 1. \href{https://doi.org/10.1038/s41598-020-76425-3}{Link}

Du J (2022). \emph{CNVds: Analyzing Human CNVs Based on Dosage Sensitivity Scores}.
Unpublished. \href{https://github.com/jenydu/CNVds}{URL}.
}
