% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDiseaseAssoc.R
\name{getDiseaseAssoc}
\alias{getDiseaseAssoc}
\title{Get Psychiatric diseases associated with CNVs}
\usage{
getDiseaseAssoc(gene_list)
}
\arguments{
\item{gene_list}{A character vector containing genes as HGNC symbols (e.g. ‘COMT’, ‘DRD3’, ‘HTR1A’).}
}
\value{
Returns a data frame containing genes and their known association
with psychiatric diseases.
}
\description{
A function that retrieves gene-disease associations from the PsyGeNet database,
based on a provided list of genes (HGNC symbols) encompassed by Copy Number Variants (CNVs).
}
\examples{
# Example 1
# Using example_CNV_call dataset available with package

# Get list of genes
gene_list <- getCNVgenes(CNV_call = sample_CNV_call)

# Get gene-disease association
gene_disease_assoc <- getDiseaseAssoc(gene_list)

# Example 2
# Produces error
gene_list_2 <- c(123, 234, 345)
gene_disease_assoc2 <- getDiseaseAssoc(gene_list2)

\dontrun{
# Example 3
# Larger dataset, runs slower
# Get list of genes
large_gene_list <- getCNVgenes(CNV_call = sample_CNV_call)

# Get gene-disease association
gene_disease_assoc3 <- getDiseaseAssoc(large_gene_list)
}

}
\references{
Akaike, H. (1973). Information theory and an extension of the maximum
likelihood principle. In \emph{Second International Symposium on Information
Theory}, New York, NY, USA, pp. 267–281. Springer Verlag. \href{https://link.springer.com/chapter/10.1007/978-1-4612-1694-0_15}{Link}

Biernacki, C., G. Celeux, and G. Govaert (2000). Assessing a mixture model for
clustering with the integrated classification likelihood. \emph{IEEE Transactions on Pattern
Analysis and Machine Intelligence} 22. \href{https://hal.inria.fr/inria-00073163/document}{Link}

Schwarz, G. (1978). Estimating the dimension of a model. \emph{The Annals of Statistics} 6, 461–464.
\href{https://projecteuclid.org/euclid.aos/1176344136}{Link}.

Yaqing, S. (2012). MBCluster.Seq: Model-Based Clustering for RNA-seq
Data. R package version 1.0.
\href{https://CRAN.R-project.org/package=MBCluster.Seq}{Link}.
}
