% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCNVgeneImpact.R
\name{plotCNVgeneImpact}
\alias{plotCNVgeneImpact}
\title{Plot Number genic vs. non-genic CNVs}
\usage{
plotCNVgeneImpact(genic_CNV_count, total_CNV_count)
}
\arguments{
\item{genic_CNV_count}{A positive integer indicating the number of CNVs
that contain 1 or more genes.}

\item{total_CNV_count}{A positive integer indicating the total number of
unique CNVs from the input table.}
}
\value{
Returns a piechart showing the number of CNVs that
contain 1 or more genes vs. CNVs that contain no genes.
}
\description{
Given the number of CNVs that contain 1 or more genes and the total number of CNVs
from the input table, plot a piechart demonstrating the distribution of CNVs that
contain 1 or more genes vs. CNVs that contain no genes. This function is called
from within getCNVgenes().
}
\examples{
# Example 1
# Using example_CNV_call dataset available with package
cnv_data <- example_CNV_call
colnames(cnv_data) <- c("chromosome_name", "start_position", "end_position", "type")
cnv_data$chromosome_name <- gsub("^chr", "", cnv_data$chromosome_name)

# Get list of genes
gene_list <- getCNVgenes(CNV_call = cnv_data, show_piechart = TRUE)


\dontrun{
# Example 2
# Obtain an external sample RNAseq dataset
# Need to download package using install.packages("MBCluster.Seq")
library(MBCluster.Seq)
data("Count")
dim(Count)

# Calculate information criteria value
InfCriteriaResults <- InfCriteriaCalculation(loglikelihood = -5080,
                                             nClusters = 2,
                                             dimensionality = ncol(Count),
                                             observations = nrow(Count),
                                             probability = c(0.5, 0.5))
InfCriteriaResults$BICresults
}

}
