% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCNVsize.R
\name{plotCNVsize}
\alias{plotCNVsize}
\title{Plot distribution of CNV sizes}
\usage{
plotCNVsize(CNV_call, chromosome_number = NULL)
}
\arguments{
\item{CNV_call}{A dataframe of CNVs containing 4 columns: the chromosome number
on which it is found, the start position, the end position, and
the type (either DUP or DEL).}

\item{chromosome_number}{A string value indicating the chromosome name
for which to plot CNV size distribution.
Can be 1-22, 'X', or 'Y'.}
}
\value{
Returns a histogram of the CNV size distribution, separated by DEL and DUP.
}
\description{
Given a CNV call data frame, plot the distribution of CNV sizes (# base pairs),
separated by the number of deletions (DEL) vs. number of duplications (DUP).
Optionally, plots the size distribution for a specified chromosome.
}
\examples{
# Example 1
# Using example_CNV_call dataset available with package
# Plot distribution of CNV sizes
plotCNVsize(CNV_call = sample_CNV_call)

# Example 2
# Produces error
cnv_data_modified <- sample_CNV_call[, 1:3]
plotCNVsize(CNV_call = cnv_data_modified)

# Example 3
# Specifying chromosome number
plotCNVsize(CNV_call = sample_CNV_call, chromosome_number = "22")


\dontrun{
# Example 4
# Larger dataset, runs slower
plotCNVsize(CNV_call = large_CNV_call)
}

}
\references{
Akaike, H. (1973). Information theory and an extension of the maximum
likelihood principle. In \emph{Second International Symposium on Information
Theory}, New York, NY, USA, pp. 267–281. Springer Verlag. \href{https://link.springer.com/chapter/10.1007/978-1-4612-1694-0_15}{Link}

Biernacki, C., G. Celeux, and G. Govaert (2000). Assessing a mixture model for
clustering with the integrated classification likelihood. \emph{IEEE Transactions on Pattern
Analysis and Machine Intelligence} 22. \href{https://hal.inria.fr/inria-00073163/document}{Link}

Schwarz, G. (1978). Estimating the dimension of a model. \emph{The Annals of Statistics} 6, 461–464.
\href{https://projecteuclid.org/euclid.aos/1176344136}{Link}.

Yaqing, S. (2012). MBCluster.Seq: Model-Based Clustering for RNA-seq
Data. R package version 1.0.
\href{https://CRAN.R-project.org/package=MBCluster.Seq}{Link}.
}
