% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDiseaseCloud.R
\name{plotDiseaseCloud}
\alias{plotDiseaseCloud}
\title{Produce a wordcloud of the most common diseases}
\usage{
plotDiseaseCloud(disease_assoc_tbl)
}
\arguments{
\item{disease_assoc_tbl}{A data frame containing gene-disease associations. Must
contain a column called DiseaseName.}
}
\value{
Returns a wordcloud plot where the size of the word represents its
frequency in a gene-disease association table.
}
\description{
Given a data frame of gene-disease associations, produces a wordcloud showing the
most frequent disease names associated with a list of genes found in Copy Number Variants (CNVs).
The size of the word represents its frequency in the gene-disease
association results.
}
\examples{
# Example 1
# Using example_CNV_call dataset available with package
# Get list of genes
gene_list <- getCNVgenes(CNV_call = sample_CNV_call)

# Get gene-disease association
gene_disease_assoc <- getDiseaseAssoc(gene_list)

# Plot wordcloud of disease/disorder names
plotDiseaseCloud(gene_disease_assoc)

# Example 2
# Produces error
plotDiseaseCloud(c("ABC", "DEF", "GHI"))

\dontrun{
# Example 3
# Larger dataset, runs slower
# Get list of genes
large_gene_list <- getCNVgenes(CNV_call = sample_CNV_call)

# Get gene-disease association
gene_disease_assoc2 <- getDiseaseAssoc(large_gene_list)

# Plot wordcloud of disease/disorder names
plotDiseaseCloud(gene_disease_assoc2)
}

}
\references{
Akaike, H. (1973). Information theory and an extension of the maximum
likelihood principle. In \emph{Second International Symposium on Information
Theory}, New York, NY, USA, pp. 267–281. Springer Verlag. \href{https://link.springer.com/chapter/10.1007/978-1-4612-1694-0_15}{Link}

Biernacki, C., G. Celeux, and G. Govaert (2000). Assessing a mixture model for
clustering with the integrated classification likelihood. \emph{IEEE Transactions on Pattern
Analysis and Machine Intelligence} 22. \href{https://hal.inria.fr/inria-00073163/document}{Link}

Schwarz, G. (1978). Estimating the dimension of a model. \emph{The Annals of Statistics} 6, 461–464.
\href{https://projecteuclid.org/euclid.aos/1176344136}{Link}.

Yaqing, S. (2012). MBCluster.Seq: Model-Based Clustering for RNA-seq
Data. R package version 1.0.
\href{https://CRAN.R-project.org/package=MBCluster.Seq}{Link}.
}
