% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPsychCNVassoc.R
\name{runPsychCNVassoc}
\alias{runPsychCNVassoc}
\title{Launch Shiny App for psychCNVassoc package}
\usage{
runPsychCNVassoc()
}
\value{
No return value but open up a Shiny page.
}
\description{
A function that launches the Shiny app for psychCNVassoc.
The code for the Shiny app is in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{

psychCNVassoc::runPsychCNVassoc()
}

}
\references{
Silva, A. (2022) TestingPackage: An Example R Package For BCB410H.  Unpublished. URL
https://github.com/anjalisilva/TestingPackage.
}
