% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_CNV_call}
\alias{sample_CNV_call}
\title{CNV position data from a CNV call}
\format{
A data frame with columns:
\describe{
\item{chromosome_name}{Chromosome name on which the CNV is found.}
\item{start_position}{Start position (base) of the CNV.}
\item{end_position}{End position (base) of the CNV.}
\item{type}{Type of the CNV, either DUP (duplication) or DEL (deletion).}
}
}
\source{
Gurbich, T.A., Ilinsky, V.V. (2020). ClassifyCNV: a tool for clinical
annotation of copy-number variants. Sci Rep 10, 20375. \href{https://doi.org/10.1038/s41598-020-76425-3}{Link}
}
\usage{
sample_CNV_call
}
\description{
A data frame containing a sample CNV call. Obtained from a sample CNV call
dataset named ACMG_examples.hg19.bed from the ClassifyCNV package in the /Examples subdirectory.
This dataset contains a subset of 21 CNV calls described by 4 fields: chromosome name, start position,
end position, and type (deletion or duplication). The CNV calling was done on Homo sapiens genotype
data using genome build hg19 and contain autosomes 1-5, 6, 11-3, 15, 17-19, 22, and sex chromosome X.
}
\examples{
\dontrun{
 sample_CNV_call
}
}
\keyword{datasets}
