% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCNVgenes.R
\name{getCNVgenes}
\alias{getCNVgenes}
\title{Get genes contained within CNVs}
\usage{
getCNVgenes(
  CNV_call,
  chromosome_number = NULL,
  reference_genome = "GRCh38",
  show_piechart = FALSE
)
}
\arguments{
\item{CNV_call}{A dataframe of CNVs containing 4 columns: the chromosome number
on which it is found, the start position, the end position, and
the type (either DUP or DEL).}

\item{chromosome_number}{A string value indicating the chromosome name
on which to identify genes contained within the CNVs.
Can be 1-22, 'X', or 'Y'.}

\item{reference_genome}{A string value, either 'GRCh37' or 'GRCh38', indicating
the genome to compare. Default is 'GRCh38'.}

\item{show_piechart}{A boolean value indicating whether to show the distribution
of genic vs. non-genic CNVs.}
}
\value{
Returns a character vector of genes that are contained within
the input list of CNVs.
}
\description{
A function that produces a list of genes that are contained within Copy Number
Variants(CNVs) from the provided table of CNVs containing the chromosome number,
start position, end position, and type (duplication or deletion). User can also
specify the chromosome number for which they want results, and the reference
genome they want to use to find genes within the CNVs. Optionally, the function
outputs a graph showing the distribution of genic vs. non-genic CNVs.
}
\examples{
# Example 1
# Using sample_CNV_call dataset available with package
# Get list of genes
gene_list <- getCNVgenes(CNV_call = sample_CNV_call)

# Example 2
# Causes an error
cnv_data_modified <- sample_CNV_call[, 1:3]

gene_list2 <- getCNVgenes(CNV_call = cnv_data_modified)


# Example 3
# Produces piechart plot
gene_list3 <- getCNVgenes(CNV_call = sample_CNV_call, show_piechart = TRUE)

# Example 4
# Specifying chromosome number and reference genome
# Get list of genes for chromosome 22 using the GRCh37 reference genome
gene_list4 <- getCNVgenes(CNV_call = sample_CNV_call, chromosome_number = "22", reference_genome = "GRCh37")


\dontrun{
# Example 5
# Larger dataset, runs slower
large_gene_list <- getCNVgenes(CNV_call = large_CNV_call)
}

}
\references{
Akaike, H. (1973). Information theory and an extension of the maximum
likelihood principle. In \emph{Second International Symposium on Information
Theory}, New York, NY, USA, pp. 267–281. Springer Verlag. \href{https://link.springer.com/chapter/10.1007/978-1-4612-1694-0_15}{Link}

Biernacki, C., G. Celeux, and G. Govaert (2000). Assessing a mixture model for
clustering with the integrated classification likelihood. \emph{IEEE Transactions on Pattern
Analysis and Machine Intelligence} 22. \href{https://hal.inria.fr/inria-00073163/document}{Link}

Schwarz, G. (1978). Estimating the dimension of a model. \emph{The Annals of Statistics} 6, 461–464.
\href{https://projecteuclid.org/euclid.aos/1176344136}{Link}.

Yaqing, S. (2012). MBCluster.Seq: Model-Based Clustering for RNA-seq
Data. R package version 1.0.
\href{https://CRAN.R-project.org/package=MBCluster.Seq}{Link}.
}
