% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDiseaseAssoc.R
\name{getDiseaseAssoc}
\alias{getDiseaseAssoc}
\title{Get Psychiatric diseases associated with CNVs}
\usage{
getDiseaseAssoc(gene_list)
}
\arguments{
\item{gene_list}{A character vector containing genes as HGNC symbols (e.g. ‘COMT’, ‘DRD3’, ‘HTR1A’).}
}
\value{
Returns a data frame containing genes and their known association
with psychiatric diseases.
}
\description{
A function that retrieves gene-disease associations from the PsyGeNet database,
based on a provided list of genes (HGNC symbols) encompassed by Copy Number Variants (CNVs).
}
\examples{
#' # Example 1
# Using example_CNV_call dataset available with package

# Get list of genes
result <- psychCNVassoc::getCNVgenes(CNV_call = sample_CNV_call)
gene_list <- result$gene_list

# Get gene-disease association
gene_disease_assoc <- psychCNVassoc::getDiseaseAssoc(gene_list)


\dontrun{
# Example 2
# Larger dataset, runs slower
# Get list of genes
result2 <- psychCNVassoc::getCNVgenes(CNV_call = sample_CNV_call)
large_gene_list <- result2$gene_list

# Get gene-disease association
gene_disease_assoc2 <- psychCNVassoc::getDiseaseAssoc(large_gene_list)
}

}
\references{
Gutierrez-Sacristan A, Hernandez-Ferrer C, Gonzalez J, Furlong L (2023).
\emph{psygenet2r: psygenet2r - An R package for querying PsyGeNET and to perform
comorbidity studies in psychiatric disorders}. doi:10.18129/B9.bioc.psygenet2r
\href{https://doi.org/10.18129/B9.bioc.psygenet2r}{DOI}, R package version 1.33.5,
\href{https://bioconductor.org/packages/psygenet2r}{Link}.
}
