% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCNVsize.R
\name{plotCNVsize}
\alias{plotCNVsize}
\title{Plot distribution of CNV sizes}
\usage{
plotCNVsize(CNV_call, chromosome_number = NULL)
}
\arguments{
\item{CNV_call}{A dataframe of CNVs containing 4 columns: the chromosome number
on which it is found, the start position, the end position, and
the type (either DUP or DEL).}

\item{chromosome_number}{A string value indicating the chromosome name
for which to plot CNV size distribution.
Can be 1-22, 'X', or 'Y'.}
}
\value{
Returns a histogram of the CNV size distribution, separated by DEL and DUP.
}
\description{
Given a CNV call data frame, plot the distribution of CNV sizes (# base pairs),
separated by the number of deletions (DEL) vs. number of duplications (DUP).
Optionally, plots the size distribution for a specified chromosome.
}
\examples{

\dontrun{

# Example 1
# Using example_CNV_call dataset available with package
# Plot distribution of CNV sizes
plotCNVsize(CNV_call = sample_CNV_call)

# Example 2
# Specifying chromosome number
plotCNVsize(CNV_call = sample_CNV_call, chromosome_number = "22")

# Example 3
# Larger dataset, runs slower
plotCNVsize(CNV_call = large_CNV_call)
}

}
\references{
Wickham H (2016). \emph{ggplot2: Elegant Graphics for Data Analysis}.
Springer-Verlag New York. \href{https://ggplot2.tidyverse.org}{Link}.

Du J (2023). \emph{CNVRS-study}. Unpublished. \href{https://github.com/jenydu/CNVRS-study}{URL}.

Gurbich T, Ilinsky V (2020). \emph{ClassifyCNV: A tool for clinical annotation of copy-number variants}.
Scientific Reports, 10(1), Article 1. \href{https://doi.org/10.1038/s41598-020-76425-3}{Link}
}
