% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDiseaseCloud.R
\name{plotDiseaseCloud}
\alias{plotDiseaseCloud}
\title{Produce a wordcloud of the most common diseases}
\usage{
plotDiseaseCloud(disease_assoc_tbl)
}
\arguments{
\item{disease_assoc_tbl}{A data frame containing gene-disease associations. Must
contain a column called DiseaseName.}
}
\value{
Returns a wordcloud plot where the size of the word represents its
frequency in a gene-disease association table.
}
\description{
Given a data frame of gene-disease associations, produces a wordcloud showing the
most frequent disease names associated with a list of genes found in Copy Number Variants (CNVs).
The size of the word represents its frequency in the gene-disease
association results.
}
\examples{
# Example 1
# Using example_CNV_call dataset available with package
# Get list of genes
gene_list <- getCNVgenes(CNV_call = sample_CNV_call)

# Get gene-disease association
gene_disease_assoc <- getDiseaseAssoc(gene_list)

# Plot wordcloud of disease/disorder names
plotDiseaseCloud(gene_disease_assoc)

# Example 2
# Produces error
plotDiseaseCloud(c("ABC", "DEF", "GHI"))

\dontrun{
# Example 3
# Larger dataset, runs slower
# Get list of genes
large_gene_list <- getCNVgenes(CNV_call = sample_CNV_call)

# Get gene-disease association
gene_disease_assoc2 <- getDiseaseAssoc(large_gene_list)

# Plot wordcloud of disease/disorder names
plotDiseaseCloud(gene_disease_assoc2)
}

}
\references{
Feinerer I, Hornik K (2023). \emph{tm: Text Mining Package}.
R package version 0.7-11, \href{https://CRAN.R-project.org/package=tm}{Link}.

Lang D, Chien G (2018). \emph{wordcloud2: Create Word Cloud by 'htmlwidget'}.
R package version 0.2.1, \href{https://CRAN.R-project.org/package=wordcloud2}{Link}.

Rul CVd (2019). \emph{How to Generate Word Clouds in R}. Medium.
\href{https://towardsdatascience.com/create-a-word-cloud-with-r-bde3e7422e8a}{Link}.

Girdher H (2023). \emph{TDM (Term Document Matrix) and DTM (Document Term Matrix)}. Analytics Vidhya.
\href{https://medium.com/analytics-vidhya/tdm-term-document-matrix-and-dtm-document-term-matrix-8b07c58957e2}{Link}.
}
